devtools/client/performance-new/test/chrome/test_perf-state-04.html

<!DOCTYPE HTML>
<html>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<head>
  <meta charset="utf-8">
  <title>Perf component test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
  <div id="container"></div>

  <pre id="test">
    <script src="head.js" type="application/javascript"></script>
    <script type="application/javascript">
      "use strict";

      /**
       * Test that the profiler gets disabled during private browsing.
       */
      addPerfTest(async () => {
        const {
          perfFrontMock,
          getRecordingState,
          mountComponent,
        } = createPerfComponent();

        perfFrontMock.mockIsLocked = true;

        mountComponent();

        is(getRecordingState(), "not-yet-known",
          "The component at first is in an unknown state.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "locked-by-private-browsing",
          "After talking to the actor, it's locked for private browsing.");

        perfFrontMock.mockIsLocked = false;
        perfFrontMock.emit("profile-unlocked-from-private-browsing");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "available-to-record",
          "After the profiler is unlocked, it's available to record.");

        document.querySelector("button").click();
        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "recording",
          "The actor has started its recording");

        perfFrontMock.mockIsLocked = true;
        perfFrontMock.emit("profile-locked-by-private-browsing");
        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "locked-by-private-browsing",
          "The recording stops when going into private browsing mode.");
      });
    </script>
  </pre>
</body>
</html>
